(refactor) further changes to string highlighting
