+++ src/org/vismayb/mocha/view/component/TextFileEditor.java
@@ -4,13 +4,12 @@ import javafx.geometry.Insets;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.layout.*;
 import javafx.scene.text.Font;
+import javafx.scene.text.FontSmoothingType;
 import javafx.scene.text.Text;
 import org.vismayb.mocha.backend.model.EditorModel;
-import org.vismayb.mocha.backend.util.FileUtil;
-import org.vismayb.mocha.frontend.util.ColorHelperKt;
+import org.vismayb.mocha.view.util.ColorHelperKt;
 
 import java.io.File;
-import java.util.ArrayList;
 
 public class TextFileEditor extends ScrollPane {
     /** Currently the "model" */
@@ -18,20 +17,19 @@ public class TextFileEditor extends ScrollPane {
     /** Should not be made new, children should be cleared and repopulated with {@code recreateFileView()}*/
     private final VBox lineContainer = new VBox();
     private final VBox gutter = new VBox();
-    private final ArrayList<String> lines;
 
     public TextFileEditor(File file) {
         super();
 
-        lines = FileUtil.getAllLines(file);
 
         // Initialize the buffer before because we don't want to do extra checks in other methods.
         model = new EditorModel(file);
 
-        for(int i =  0; i < lines.size(); i++) {
+        for(int i =  0; i < model.getLines().size(); i++) {
             var t = new Text(Integer.toString(i + 1));
             t.setFont(Font.font("JetBrains Mono", 15));
             t.setFill(ColorHelperKt.generateColor(95, 99, 102));
+            t.setFontSmoothingType(FontSmoothingType.LCD);
             gutter.getChildren().add(t);
         }
 
@@ -58,8 +56,7 @@ public class TextFileEditor extends ScrollPane {
         System.out.println(model.getLines().get(0));
         for(int i = 0; i <= model.getLines().size(); i++) {
             var tokensInLine = model.getTokensByLineNumber(i);
-            System.out.println(model.getLines().get(i));
-            EditorLine line = new EditorLine(model.getLines().get(i));
+            EditorLine line = new EditorLine(model.getLines().get(i), tokensInLine, i + 1);
             lineContainer.getChildren().add(line);
         }
     }
